requirements.txt
flet==0.25.2
numpy==1.22.0
pandas==2.1.1
fuzzywuzzy==0.18.0
reportlab==4.0.5
arabic-reshaper
python-bidi
scikit-learn==1.3.2





















